scripts/requirements.txt
pymysql==1.1.0
minio==7.2.0
openpyxl==3.1.2
pandas==2.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.46
Pillow==12.1.1
